emorobot/requirements.txt
Django>=2.2.5
djangorestframework>=3.10.3
paho-mqtt>=1.4.0
tensorflow==1.13.1
pandas
librosa
numpy